+++ CHANGELOG.md
@@ -1,4 +1,14 @@
 ### Added
+- Checkbox & Checkbox group component [#75]
+- Radio component [#75]
+- Select component [#75]
+- Basic Cards [#75]
+- Section header component [#75]
+- Question card component [#75]
+- Tabs component [#75]
+- Template Overview - blocked out page for component builder assistance [#75]
+- Template Edit Section  - blocked out page for component builder assistance [#75]
+- Template Edit Question - blocked out page for component builder assistance [#75]
 - Fixed type bugs in the template history page which was breaking the build.
 - Added csrf token to header for /login and /signup page and error handling to refresh csrf and auth tokens [#116]
 - Added cypress for functional tests. Added a functional test to check for HTTP-only auth cookies [#116]
@@ -12,6 +22,8 @@
 - Added the initial version of our custom ReMirror text editor [#74].
 
 ### Updates
+- Button styles [#75]
+- Text input styles [#75]
 - Made updates to change 'name' to 'displayName' for affiliations, in response to backend changes [#137]
 - Updated endpoints for 'signin' and 'signup' to be 'apollo-signin' and 'apollo-signup' [#99]